Added Difficulty Setting, Fixed Home Screen Bug
Added a difficulty option to the game, which now displays on the left of the screen. Difficulty affects the spawn frequency of notes as it allows the Peak detect algorithm to detect peaks more frequently. The error shown when moving the mouse before the home object has been initialised was fixed by implementing a check that home was not null before attempting to use the object's method in the mouseClicked() function.